Revert "Release 0.4.0"
