+++ .bumpversion.cfg
@@ -6,7 +6,7 @@ serialize = {major}.{minor}.{patch}-{release_type}{iteration}
 tag = True
 commit = True
 message = Cut New Release: {current_version} → {new_version}
-current_version = 0.3.6-dev1
+current_version = 0.3.6
 
 [bumpversion:part:release_type]
 optional_value = stable
