+++ .circleci/config.yml
@@ -37,6 +37,7 @@ PR_branches_only: &PR_branches_only
       ignore:
         - /^\d+\.\d+.*/
 
+
 ################################################################################
 #                                                                              #
 # Custom Executor definitions.                                                 #
@@ -66,17 +67,6 @@ commands:
   # ================= #
   # CI Setup Commands #
   # ================= #
-
-  setup-job:
-    description: |
-      Attach the workspace and load ENV variables.
-      Additionally skips the job this is used in, if the commit message is invalid
-      (unless the workflow executing the job is running on "master" or "dev").
-    steps:
-      - attach_workspace:
-          at: "/home/circleci"
-      - export_env_vars
-      - skip_bad_commit_format
 
   export_env_vars:
     description: Export ENV variables used by our script.
@@ -124,12 +114,12 @@ commands:
             make install-dev
 
 
-  skip_bad_commit_format:
+  validate_commit_format:
     description: Validate the first line of the commit message against a REGEX.
     steps:
       - run:
           name: Validate the commit's title description.
-          command: python3 ${CI_SCRIPTS_DIR}/skip-on-bad-commit-message.py
+          command: python3 ${CI_SCRIPTS_DIR}/validate-commit-message.py
 
   # ================================== #
   # Linter and test execution commands #
@@ -181,6 +171,14 @@ commands:
 
 jobs:
 
+  validate-commit-message:
+    executor: default-executor
+    steps:
+      - attach_workspace:
+          at: "/home/circleci"
+      - export_env_vars
+      - validate_commit_format
+
   prep-system:
     executor: default-executor
     steps:
@@ -194,22 +192,11 @@ jobs:
 
   # Run all test suites (aka the test harness) of the repository and lint-check
   # our code base.
-  test-harness:
-    executor: default-executor
-    steps:
-      - setup-job
-      - run_test_harness
-
-  style-check-and-lint:
-    executor: default-executor
-    steps:
-      - setup-job
-      - lint_codebase
-
   lint-and-test:
     executor: default-executor
     steps:
-      - setup-job
+      - attach_workspace:
+          at: "/home/circleci"
       - lint_codebase
       - run_test_harness
 
@@ -218,7 +205,9 @@ jobs:
   bump-versions:
     executor: default-executor
     steps:
-      - setup-job
+      - attach_workspace:
+          at: "/home/circleci"
+      - export_env_vars
       - run:
           name: Add Github to known hosts.
           command: |
@@ -243,8 +232,6 @@ jobs:
     executor: default-executor
     steps:
       - checkout
-      - export_env_vars
-      - skip_bad_commit_format
       - run:
           name: Publish to pypi using flit.
           command: |
@@ -276,12 +263,7 @@ workflows:
       - prep-system:
           <<: *PR_branches_only
 
-      - test-harness:
-          context: Raiden-SP-Context
-          requires:
-            - prep-system
-
-      - style-check-and-lint:
+      - lint-and-test:
           context: Raiden-SP-Context
           requires:
             - prep-system
@@ -289,19 +271,23 @@ workflows:
       - finalize:
           requires:
             - prep-system
-            - style-check-and-lint
-            - test-harness
+            - lint-and-test
 
   Merge-Commit-Workflow:
     jobs:
       - prep-system:
           <<: *master_dev_only
 
+      - validate-commit-message:
+          context: Raiden-SP-Context
+          requires:
+            - prep-system
+
       # Run linters and test harnesses against PR branches.
       - lint-and-test:
           context: Raiden-SP-Context
           requires:
-            - prep-system
+            - validate-commit-message
 
       # Bump the minor or patch for `master` on each merge commit.
       - bump-versions:
@@ -321,6 +307,7 @@ workflows:
       - finalize:
           requires:
             - prep-system
+            - validate-commit-message
             - lint-and-test
             - bump-versions
             - tear-down
