+++ Dockerfile
@@ -4,13 +4,22 @@
 ARG PY_VERSION=3.7
 FROM python:$PY_VERSION AS cache
 
-RUN pip install raiden
+# Clone raiden repo and switch to its `develop` branch
+RUN git clone https://github.com/raiden-network/raiden /raiden
+RUN git --git-dir /raiden/.git checkout develop
+
+# Install raiden's development dependencies.
+RUN pip install -r /raiden/requirements/requirements-dev.txt
+
+# Install the raiden package
+RUN pip install ./raiden
 
 FROM python:${PY_VERSION}
 
 ARG PY_VERSION
 
 # Copy raiden repository and site-packages from build cache
+COPY --from=cache /raiden /raiden
 COPY --from=cache /usr/local/lib/python${PY_VERSION}/dist-packages /usr/local/lib/python${PY_VERSION}/dist-packages
 
 #  Copy SP folder and install.
