+++ scenario_player/__init__.py
@@ -2,4 +2,5 @@
 
 End-to-end testing tool for the ``Raiden`` test suite.
 """
-__version__ = "0.3.6-dev1"
+
+__version__ = "0.3.6"
