+++ scenario_player/utils/legacy.py
@@ -93,27 +93,6 @@ class ChainConfigType(click.ParamType):
         return name, rpc_url
 
 
-class MutuallyExclusiveOption(click.Option):
-    def __init__(self, *args, **kwargs):
-        self.mutually_exclusive = set(kwargs.pop("mutually_exclusive", []))
-        help = kwargs.get("help", "")
-        if self.mutually_exclusive:
-            ex_str = ", ".join(self.mutually_exclusive)
-            kwargs["help"] = help + (
-                " NOTE: This argument is mutually exclusive with " " arguments: [" + ex_str + "]."
-            )
-        super(MutuallyExclusiveOption, self).__init__(*args, **kwargs)
-
-    def handle_parse_result(self, ctx, opts, args):
-        if self.mutually_exclusive.intersection(opts) and self.name in opts:
-            raise click.UsageError(
-                f"Illegal usage: {self.name} is mutually exclusive with "
-                f"arguments {', '.join(self.mutually_exclusive)}."
-            )
-
-        return super(MutuallyExclusiveOption, self).handle_parse_result(ctx, opts, args)
-
-
 class HTTPExecutor(mirakuru.HTTPExecutor):
     def start(self, stdout=subprocess.PIPE, stderr=subprocess.PIPE):
         """ Merged copy paste from the inheritance chain with modified stdout/err behaviour """
