--- tests/unittests/cli/keystore/UTC--1
@@ -1,1 +1,0 @@
-{"id":"e66b03f8-4ce7-8f9b-b5d1-36dc9d5154ba","version":3,"crypto":{"cipher":"aes-128-ctr","cipherparams":{"iv":"49b8891b02604fb05884280509b82fbb"},"ciphertext":"4ce3a7059820b516746d7cc5defca17c2a54ea9dbb7e3a5411a01630b5021244","kdf":"pbkdf2","kdfparams":{"c":10240,"dklen":32,"prf":"hmac-sha256","salt":"9ef5e56c835cf59f4c9dbe6bd6a88b11056d6f01c0590c17d1cbb1653f749e7a"},"mac":"861354d5dd3efb2c6470090fb901957c53035175ababc9dc1863f5c07572d33b"},"address":"ec1fdb2d29c5689416b3f1b55a4d879fddf0e6e3","name":"","meta":"{}"}
