--- tests/unittests/cli/keystore/password
@@ -1,1 +1,0 @@
-123
