--- tests/unittests/cli/keystore/wrong_password
@@ -1,1 +1,0 @@
-123456
