--- tests/unittests/cli/scenario/join-network-scenario-J1.yaml
@@ -1,48 +1,0 @@
-version: 2
-
-settings:
-  gas_price: "fast"
-  chain: any
-  services:
-    pfs:
-      url: https://pfs-goerli.services-dev.raiden.network
-    udc:
-      enable: true
-      token:
-        deposit: true
-
-token:
-
-nodes:
-  mode: managed
-  count: 4
-
-  default_options:
-    gas-price: fast
-    environment-type: development
-    routing-mode: pfs
-    pathfinding-max-paths: 5
-    pathfinding-max-fee: 10
-
-scenario:
-  serial:
-    tasks:
-      - parallel:
-          name: "Setting up a network"
-          tasks:
-            - open_channel: {from: 0, to: 1, total_deposit: 10, expected_http_status: 201}
-            - open_channel: {from: 0, to: 2, total_deposit: 10, expected_http_status: 201}
-            - open_channel: {from: 1, to: 2, total_deposit: 10, expected_http_status: 201}
-      - serial:
-          name: "Checking the network"
-          tasks:
-            - assert: {from: 0, to: 1, total_deposit: 10, balance: 10, state: "opened"}
-            - assert: {from: 0, to: 2, total_deposit: 10, balance: 10, state: "opened"}
-            - assert: {from: 1, to: 2, total_deposit: 10, balance: 10, state: "opened"}
-      - serial:
-          name: "Node Nr. 4 joins"
-          tasks:
-            - join_network: {from: 3, funds: 100, initial_channel_target: 3, joinable_funds_target: 0.4, expected_http_status: 204}
-            - assert: {from: 3, to: 0, total_deposit: 20, balance: 20, state: "opened"}
-            - assert: {from: 3, to: 1, total_deposit: 20, balance: 20, state: "opened"}
-            - assert: {from: 3, to: 2, total_deposit: 20, balance: 20, state: "opened"}
